requirements.txt
aws_cdk.core==1.22.0
aws_cdk.aws_lambda==1.22.0
aws_cdk.aws_apigateway==1.22.0
aws_cdk.aws_sqs==1.22.0
aws_cdk.aws_sns==1.22.0
aws_cdk.aws_s3==1.22.0
aws_cdk.aws_dynamodb==1.22.0
aws_cdk.aws_appsync==1.22.0
aws_cdk.aws_cognito==1.22.0
aws_cdk.aws_cloudfront==1.22.0
aws_cdk.aws_iam==1.22.0
aws_cdk.aws_stepfunctions==1.22.0
aws_cdk.aws_lambda_event_sources==1.22.0
aws_cdk.aws_stepfunctions_tasks==1.22.0
aws_cdk.aws_s3_deployment==1.22.0